adding the category and order controllers

+++ TechXpress.Web/ViewModel/OrderViewModel.cs
@@ -5,7 +5,7 @@
         public int OrderId { get; set; }
         public string CustomerName { get; set; }
         public DateTime OrderDate { get; set; }
-        public decimal TotalAmount { get; set; }
+        public float TotalAmount { get; set; }
         public string Status { get; set; }
     }
 }
